src/main/kotlin/no/skatteetaten/aurora/mean/genie/service/ApplicationDeploymentWatcherService.kt
package no.skatteetaten.aurora.mean.genie.service

import com.fasterxml.jackson.databind.JsonNode
import com.fasterxml.jackson.databind.node.ArrayNode
import mu.KotlinLogging
import org.springframework.beans.factory.annotation.Value
import org.springframework.scheduling.annotation.Async
import org.springframework.stereotype.Service
import reactor.core.publisher.Mono

private val logger = KotlinLogging.logger {}

@Service
class ApplicationDeploymentWatcherService(
    @Value("\${integrations.operations.scope:}") val operationScopeConfiguration: String?,
    val watcher: KubernetesWatcher,
    val databaseService: DatabaseService
) {
    @Async
    fun watch() {
        val labelSelector = checkForOperationScopeLabel()
        val url = "/apis/skatteetaten.no/v1/applicationdeployments?watch=true&labelSelector=$labelSelector"
        watcher.watch(url, listOf("DELETED")) { event ->
            deleteSchemasIfExists(event)
        }
    }

    fun checkForOperationScopeLabel(): String {
        return if (operationScopeConfiguration.isNullOrEmpty()) {
            "!operationScope"
        } else {
            "operationScope=$operationScopeConfiguration"
        }
    }

    fun deleteSchemasIfExists(event: JsonNode): Mono<Void> {
        val jsonArray = event.at("/object/spec/databases") as ArrayNode
        val databases = jsonArray.map { it.textValue() }

        return if (databases.isEmpty()) {
            Mono.empty()
        } else {
            logger.debug { "Attempting to delete database schema $databases" }
            databaseService.deleteSchemaByID(databases)
                .then()
        }
    }
}
